requirements.txt
aio-pika==9.5.7
aiormq==6.9.2
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
colorama==0.4.6
coverage==7.11.3
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.121.1
h11==0.16.0
hiredis==3.3.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
multidict==6.7.0
packaging==25.0
pamqp==3.3.0
passlib==1.7.4
pluggy==1.6.0
propcache==0.4.1
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.0
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
redis==7.0.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
yarl==1.22.0
